requirements.txt
fastapi
uvicorn
gunicorn
yt-dlp
telethon
python-dotenv